refactor: centralize common utilities
