+++ .github/workflows/deploy-lambda.yml
@@ -6,6 +6,7 @@ on:
       - '.github/workflows/deploy-lambda.yml'
       - 'lambda_bot/**'
       - 'lambda_bot_twitter_scraper/**'
+      - 'common/**'
 
 jobs:
   deploy:
